refactor(platforms): merge Feishu platform modules and cleanup deprecated files
Implements Phase 2 of Issue #267:
- Move channels/platforms/feishu/ to platforms/feishu/
- Update all import paths to use new location
- Add backward compatibility re-export in channels/platforms/feishu/index.ts
- Delete deprecated feishu/message-history.ts re-export

Changes:
- Moved Feishu platform modules to platforms/feishu/
- Updated import paths in feishu-channel.ts, factory.ts, index.ts
- Updated test mocks in platform-adapter.test.ts and communication-node.test.ts
- Added @deprecated re-export for backward compatibility

Closes #267
