+++ src/channels/platforms/feishu/index.ts
@@ -1,15 +1,11 @@
 /**
- * Feishu Platform Module.
+ * Feishu Platform Module (Legacy Location).
  *
- * Exports Feishu-specific implementations of platform adapters.
+ * @deprecated Import from '../../../platforms/feishu/index.js' instead.
+ * This file is kept for backward compatibility.
+ *
+ * @see Issue #267 - Phase 2: 合并平台模块
  */
 
-// Platform Adapter
-export { FeishuPlatformAdapter, type FeishuPlatformAdapterConfig } from './feishu-adapter.js';
-
-// Sub-adapters
-export { FeishuMessageSender, type FeishuMessageSenderConfig } from './feishu-message-sender.js';
-export { FeishuFileHandler, type FeishuFileHandlerConfig } from './feishu-file-handler.js';
-
-// Card Builders
-export { buildTextContent } from './card-builders/index.js';
+// Re-export from new location for backward compatibility
+export * from '../../../platforms/feishu/index.js';
