--- src/feishu/message-history.ts
@@ -1,14 +1,0 @@
-/**
- * Message history manager re-export.
- *
- * @deprecated Import from '../core/index.js' instead.
- * This file is kept for backward compatibility.
- */
-
-export {
-  MessageHistoryManager,
-  messageHistoryManager,
-  type IMessageHistoryManager,
-  type ChatMessage,
-  type ChatHistory,
-} from '../core/message-history.js';
